fix(money): 金額尾數保留小數 🟡7
金額其他欄位早是 numeric(10,2)（支援 $63.3/公斤），但兩處還在砍小數：
- orders/purchase_orders.total_amount 是 integer + SUM ::int → 訂單總額
  跟明細加總對不起來。migration 改 numeric(10,2)（int→numeric 無損），
  移除 ::int，recalcOrderTotal 用 Number(numeric SUM)
- 預約改價表單 parseInt(newCostPrice) → 63.3 變 63。改 parseFloat
  （DB 欄位本來就是 numeric，只是前端砍掉）

驗證 rollback-tx：63.30+12.35 訂單總額存回 75.65（不是 75/76）。

+++ src/app/dashboard/price-schedule/page.tsx
@@ -104,8 +104,9 @@ export default function PriceSchedulePage() {
       headers: { 'Content-Type': 'application/json' },
       body: JSON.stringify({
         itemId: parseInt(formData.itemId),
-        newCostPrice: parseInt(formData.newCostPrice),
-        newStorePrice: formData.newStorePrice ? parseInt(formData.newStorePrice) : null,
+        // 金額保留小數（$63.3/公斤 這種），不要 parseInt 砍掉
+        newCostPrice: parseFloat(formData.newCostPrice),
+        newStorePrice: formData.newStorePrice ? parseFloat(formData.newStorePrice) : null,
         effectiveDate: formData.effectiveDate,
         source: formData.source || null,
         notes: formData.notes || null,
@@ -134,7 +135,7 @@ export default function PriceSchedulePage() {
 
   const selectedItem = items.find((i) => i.id === parseInt(formData.itemId))
   const priceDiff = selectedItem && formData.newCostPrice
-    ? parseInt(formData.newCostPrice) - selectedItem.costPrice
+    ? parseFloat(formData.newCostPrice) - selectedItem.costPrice
     : null
 
   return (
